NNTrainer: added basic output methods
