+++ src/NNTrainerGSL.hpp
@@ -16,7 +16,7 @@ class NNTrainerGSL: public NNTrainer
 {
 
 public:
-    NNTrainerGSL(NNTrainingData * tdata): NNTrainer(tdata){};
+    NNTrainerGSL(NNTrainingData * const tdata): NNTrainer(tdata){};
     void findFit(double * const fit, double * const err, double &resi_full, double &resi_noreg, double &resi_pure, const int nsteps, const bool verbose);
 };
